+++ src/Runtime/Runtime/Empty Namespaces/System.Windows.Controls/WORKINPROGRESS/AutoCompleteBox.cs
@@ -1,0 +1,25 @@
+﻿#if WORKINPROGRESS
+
+using System;
+using System.Windows.Controls.Primitives;
+using System.Windows.Documents;
+
+#if MIGRATION
+namespace System.Windows.Controls
+#else
+namespace Windows.UI.Xaml.Controls
+#endif
+{
+    public partial class AutoCompleteBox : Selector
+    {
+        virtual protected void OnPopulated(PopulatedEventArgs e)
+        {
+
+        }
+        public static DependencyProperty ItemFilterProperty
+        {
+            get; set;
+        }
+    }
+}
+#endif
