+++ src/Runtime/Runtime/Empty Namespaces/System.Windows.Controls/WORKINPROGRESS/Control.cs
@@ -1,0 +1,16 @@
+﻿#if WORKINPROGRESS
+
+using System;
+
+#if MIGRATION
+namespace System.Windows.Controls
+#else
+namespace Windows.UI.Xaml.Controls
+#endif
+{
+    public abstract partial class Control : FrameworkElement
+    {
+        public int CharacterSpacing { get; set; }
+    }
+}
+#endif
