+++ src/Runtime/Runtime/Empty Namespaces/System.Windows.Controls/WORKINPROGRESS/PopulatedEventArgs.cs
@@ -1,0 +1,19 @@
+﻿#if WORKINPROGRESS
+
+using System;
+using System.Windows.Documents;
+
+#if MIGRATION
+namespace System.Windows.Controls
+#else
+namespace Windows.UI.Xaml.Controls
+#endif
+{
+    public partial class PopulatedEventArgs
+    {
+
+    }
+
+    public delegate void RoutedPropertyChangingEventHandler<T>(Object sender, RoutedPropertyChangingEventArgs<T> e);
+}
+#endif
