+++ src/Runtime/Runtime/Empty Namespaces/System.Windows.Controls/WORKINPROGRESS/WebBrowser.cs
@@ -1,0 +1,26 @@
+﻿#if WORKINPROGRESS
+
+using System;
+
+#if MIGRATION
+namespace System.Windows.Controls
+#else
+namespace Windows.UI.Xaml.Controls
+#endif
+{
+    public partial class WebBrowser
+    {
+        public string SaveToString()
+        {
+            return "";
+        }
+
+        public static DependencyProperty SourceProperty
+        {
+            get; set;
+        }
+
+        public System.Uri Source { get; set; }
+    }
+}
+#endif
