+++ src/Runtime/Runtime/Empty Namespaces/System.Windows.Data/WORKINPROGRESS/Binding.cs
@@ -1,11 +1,12 @@
 ﻿#if WORKINPROGRESS
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
+using System;
+
+#if MIGRATION
 namespace System.Windows.Data
+#else
+namespace Windows.UI.Xaml.Data
+#endif
 {
     public partial class Binding : BindingBase
     {
