+++ src/Runtime/Runtime/Empty Namespaces/System.Windows.Documents/WORKINPROGRESS/FontSource.cs
@@ -1,11 +1,12 @@
 ﻿#if WORKINPROGRESS
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
+using System;
+
+#if MIGRATION
 namespace System.Windows.Documents
+#else
+namespace Windows.UI.Xaml.Documents
+#endif
 {
     public class FontSource
     {
