+++ src/Runtime/Runtime/Empty Namespaces/System.Windows/WORKINPROGRESS/Window.cs
@@ -1,0 +1,20 @@
+﻿#if WORKINPROGRESS
+
+using System;
+
+#if MIGRATION
+namespace System.Windows
+#else
+namespace Windows.UI.Xaml
+#endif
+{
+    public partial class Window
+    {
+        public string Title { get; set; }
+        public bool Install()
+        {
+            return true;
+        }
+    }
+}
+#endif
